backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
httpx>=0.24.1
pydantic>=2.0.0
starlette>=0.27.0
